Unified scanner's dialog with inventory's newItem
The dialog in the scanner now is the same exact dialog as the one in the inventory, when you tap the + FAB. The barcode field just gets pre-filled in, when you use the scanner.
Also added a snackbar to alert the user when the newItem dialog wasn't filled in correctly.